move coin value checking from utxo list to avalanche dialog
This make the warning show also when building a proof from coin details file, when previouly the check was only done for coins selected in the UTXO list.
